Issue #7: Move settings into folder.
