+++ mysitedj2_0/settings/travis.py
@@ -1,7 +1,7 @@
 # coding: utf-8
 """Settings for Travis."""
 
-from mysitedj2_0.settings import * # noqa
+from mysitedj2_0.settings.base import * # noqa
 
 
 SECRET_KEY = "tlt#^9^3gg-jqm+qr__sp58$&k*8u92hyz2r1&c2iziay@on9m"
